fix import hierarchy for browser use
