+++ src/lmql/__init__.py
@@ -25,7 +25,6 @@ from lmql.runtime.model_registry import LMQLModelRegistry
 from lmql.runtime.output_writer import headless, printing, silent, stream
 from lmql.runtime.interpreter import LMQLResult
 from lmql.models.model import model
-from lmql.models.lmtp.lmtp_serve import serve
 
 model_registry = LMQLModelRegistry
 
@@ -150,3 +149,8 @@ def main(query_fct, **kwargs):
     """
     import asyncio
     return asyncio.run(query_fct(**kwargs))
+
+def serve(*args, **kwargs):
+    assert not "LMQL_BROWSER" in os.environ, "lmql.serve is not available in the browser distribution of LMQL."
+    from lmql.models.lmtp.lmtp_serve import serve
+    return serve(*args, **kwargs)
